move components to new project structure
