+++ src/atoms/avatar/styles.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const Image = styled.img`
+  border: 1px solid var(--default, hsl(0, 0%, 16%));
+  border-radius: 999px;
+  width: 30px;
+  height: 30px;
+  object-fit: cover;
+  overflow: hidden;
+`;
