+++ src/atoms/checkbox/styles.tsx
@@ -1,20 +1,18 @@
-import React, { HTMLAttributes } from 'react';
 import styled from 'styled-components';
 
 interface CheckboxErrorProps {
   error?: string;
 }
-
 interface CheckboxWrapperProps extends CheckboxErrorProps {
   disabled?: boolean;
 }
 
-const CheckboxWrapper = styled.label<CheckboxWrapperProps>`
+export const CheckboxWrapper = styled.label<CheckboxWrapperProps>`
   display: inline-block;
   position: relative;
   height: 24px;
   padding-left: 0;
-  cursor: ${(props) => (props.disabled ? 'auto' : 'pointer')};
+  cursor: ${props => (props.disabled ? 'auto' : 'pointer')};
   -webkit-user-select: none;
   -moz-user-select: none;
   -ms-user-select: none;
@@ -25,24 +23,24 @@ const CheckboxWrapper = styled.label<CheckboxWrapperProps>`
   }
 
   span {
-    color: ${(props) =>
+    color: ${props =>
       props.error
         ? 'var(--red, hsl(354, 83%, 64%))'
         : props.disabled
         ? 'var(--grey, hsl(0, 0%, 85%))'
         : null};
-    border-color: ${(props) =>
+    border-color: ${props =>
       props.error ? 'var(--red, hsl(354, 83%, 64%))' : null};
   }
 `;
 
-const Label = styled.span`
+export const CheckboxLabel = styled.span`
   padding-left: 30px;
   line-height: 24px;
   vertical-align: middle;
 `;
 
-const Input = styled.input`
+export const CheckboxInput = styled.input`
   position: absolute;
   opacity: 0;
   cursor: pointer;
@@ -76,12 +74,12 @@ const Input = styled.input`
   }
 `;
 
-const Checkmark = styled.span<CheckboxErrorProps>`
+export const Checkmark = styled.span<CheckboxErrorProps>`
   position: absolute;
   top: 0;
   left: 0;
   border: 2px solid
-    ${(props) =>
+    ${props =>
       props.error
         ? 'var(--red, hsl(354, 83%, 64%))'
         : 'var(--grey, hsl(0, 0%, 85%))'};
@@ -105,45 +103,3 @@ const Checkmark = styled.span<CheckboxErrorProps>`
     transform: rotate(45deg);
   }
 `;
-
-interface Props extends HTMLAttributes<HTMLDivElement> {
-  value: string;
-  label?: string;
-  checked?: boolean;
-  onChange?: () => {};
-  error?: string;
-  disabled?: boolean;
-}
-
-const Checkbox = (props: Props) => {
-  const {
-    label,
-    value,
-    checked = false,
-    onChange = () => {},
-    error,
-    disabled = false,
-    style,
-    className = '',
-  } = props;
-  return (
-    <CheckboxWrapper
-      error={error}
-      disabled={disabled}
-      style={style}
-      className={className}
-    >
-      <Label>{label}</Label>
-      <Input
-        type="checkbox"
-        name={value}
-        checked={checked}
-        onChange={onChange}
-        disabled={disabled}
-      />
-      <Checkmark />
-    </CheckboxWrapper>
-  );
-};
-
-export default Checkbox;
