+++ src/atoms/tag/types.ts
@@ -1,0 +1,1 @@
+export type TagColor = 'primary' | 'orange' | 'danger' | 'info' | 'outline';
