+++ src/index.ts
@@ -6,8 +6,8 @@ export { default as TaikaiLogo } from './ions/taikai-logo';
 
 // Atoms
 export { default as Avatar } from './atoms/avatar';
-export { default as Checkbox } from './atoms/checkbox';
-export { default as Tag } from './atoms/tag';
-export { default as Spinner } from './atoms/spinner';
-export { default as Icon } from './atoms/icon';
 export { default as Button } from './atoms/button';
+export { default as Checkbox } from './atoms/checkbox';
+export { default as Icon } from './atoms/icon';
+export { default as Spinner } from './atoms/spinner';
+export { default as Tag } from './atoms/tag';
