+++ test/atoms/avatar.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
-import { Avatar } from '../src';
+import { Avatar } from '../../src';
 
 describe('Avatar', () => {
   it('renders', () => {
